Start clean data
Started to work out how I shall clean this dataset
so I can start analyse it, filled in dates that
was missing since I´m working with actual facts
that happened.

Next step is to drop columns that was missing info
since those won´t be necessary in my project.