Created databases
First I´ve changed my .xlsx file to .csv with a
python code, then I´ve imported both my .csv files
to separate .db files.

With SQLite3 Editor it´s easy to see the tables
and change types on the columns.
All columns are generated to be TEXT but changed
some to PK, INT, DECIMAL and DATE.
Add a REFERENCES type to one of the tables so they
are linked up.