Multiselect on map_page
Added a filter so that the user can choose one or
many fates to be shown on the map so they can find
the fates that they are interested in.

also fixed that the column AverageDayPerShip now
are converted into an integer.

Changed the explaining of all the fates to more
distinct information.